chore: move the administration controllers to their own package
