+++ backend/api/src/main/kotlin/io/tolgee/api/v2/controllers/administration/AdministrationController.kt
@@ -1,7 +1,8 @@
-package io.tolgee.api.v2.controllers
+package io.tolgee.api.v2.controllers.administration
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
+import io.tolgee.api.v2.controllers.IController
 import io.tolgee.constants.Message
 import io.tolgee.dtos.queryResults.organization.OrganizationView
 import io.tolgee.exceptions.BadRequestException
@@ -45,23 +46,23 @@ import org.springframework.web.bind.annotation.RestController
 )
 @OpenApiSelfHostedExtension
 class AdministrationController(
-  private val organizationService: OrganizationService,
-  private val pagedOrganizationResourcesAssembler: PagedResourcesAssembler<OrganizationView>,
-  private val organizationModelAssembler: OrganizationModelAssembler,
-  private val authenticationFacade: AuthenticationFacade,
-  private val userAccountService: UserAccountService,
-  private val pagedResourcesAssembler: PagedResourcesAssembler<UserAccount>,
-  private val userAccountModelAssembler: UserAccountModelAssembler,
-  private val jwtService: JwtService,
+    private val organizationService: OrganizationService,
+    private val pagedOrganizationResourcesAssembler: PagedResourcesAssembler<OrganizationView>,
+    private val organizationModelAssembler: OrganizationModelAssembler,
+    private val authenticationFacade: AuthenticationFacade,
+    private val userAccountService: UserAccountService,
+    private val pagedResourcesAssembler: PagedResourcesAssembler<UserAccount>,
+    private val userAccountModelAssembler: UserAccountModelAssembler,
+    private val jwtService: JwtService,
 ) : IController {
   @GetMapping(value = ["/organizations"])
   @Operation(summary = "Get all server organizations")
   @RequiresSuperAuthentication
   fun getOrganizations(
-    @ParameterObject
+      @ParameterObject
     @SortDefault(sort = ["name"])
     pageable: Pageable,
-    search: String? = null,
+      search: String? = null,
   ): PagedModel<OrganizationModel> {
     val organizations =
       organizationService.findAllPaged(
@@ -76,10 +77,10 @@ class AdministrationController(
   @Operation(summary = "Get all server users")
   @RequiresSuperAuthentication
   fun getUsers(
-    @ParameterObject
+      @ParameterObject
     @SortDefault(sort = ["name"])
     pageable: Pageable,
-    search: String? = null,
+      search: String? = null,
   ): PagedModel<UserAccountModel> {
     val users = userAccountService.findAllWithDisabledPaged(pageable, search)
     return pagedResourcesAssembler.toModel(users, userAccountModelAssembler)
@@ -89,7 +90,7 @@ class AdministrationController(
   @Operation(summary = "Delete user")
   @RequiresSuperAuthentication
   fun deleteUser(
-    @PathVariable userId: Long,
+      @PathVariable userId: Long,
   ) {
     if (userId == authenticationFacade.authenticatedUser.id) {
       throw BadRequestException(Message.CANNOT_DELETE_YOUR_OWN_ACCOUNT)
@@ -107,7 +108,7 @@ class AdministrationController(
   )
   @RequiresSuperAuthentication
   fun disableUser(
-    @PathVariable userId: Long,
+      @PathVariable userId: Long,
   ) {
     if (userId == authenticationFacade.authenticatedUser.id) {
       throw BadRequestException(Message.CANNOT_DISABLE_YOUR_OWN_ACCOUNT)
@@ -119,7 +120,7 @@ class AdministrationController(
   @Operation(summary = "Enable user", description = "Enables previously disabled user.")
   @RequiresSuperAuthentication
   fun enableUser(
-    @PathVariable userId: Long,
+      @PathVariable userId: Long,
   ) {
     userAccountService.enable(userId)
   }
@@ -128,8 +129,8 @@ class AdministrationController(
   @Operation(summary = "Set Role", description = "Set's the global role on the Tolgee Platform server.")
   @RequiresSuperAuthentication
   fun setRole(
-    @PathVariable userId: Long,
-    @PathVariable role: UserAccount.Role,
+      @PathVariable userId: Long,
+      @PathVariable role: UserAccount.Role,
   ) {
     val user = userAccountService.get(userId)
     user.role = role
@@ -145,7 +146,7 @@ class AdministrationController(
   )
   @RequiresSuperAuthentication
   fun generateUserToken(
-    @PathVariable userId: Long,
+      @PathVariable userId: Long,
   ): String {
     val user = userAccountService.get(userId)
     return jwtService.emitToken(user.id, true)
